app/requirements.txt
Django==5.0.3
drf_yasg==1.21.7
django_filter==24.1
django-extensions==3.2.3
django-rest-swagger==2.2.0
django_cors_headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
django_phonenumbers==1.0.1
django_phonenumber_field==7.3.0
celery==5.3.6
Pillow==10.2.0
psycopg2==2.9.9
uvicorn==0.28.0
redis==5.0.3
django-redis==5.4.0
boto3==1.34.61
minio==7.2.5
django-storages==1.14.2
cairosvg==2.7.1
Faker==24.1.0
django-db-logger==0.1.13
gunicorn==21.2.0
django_prometheus==2.3.1
asyncpg==0.29.0
django_silk==5.1.0
